* hieroglyph/hgint.h (HG_QINT): fix affecting to the flag bits 	when creating a quark with a negative.

+++ hieroglyph/hgint.h
@@ -38,7 +38,7 @@ struct _hg_bs_int_t {
 
 
 #define HG_QINT(_v_)				\
-	hg_quark_new(HG_TYPE_INT, (_v_))
+	hg_quark_new(HG_TYPE_INT, (guint32)(_v_))
 #define HG_INT(_q_)				\
 	((gint32)hg_quark_get_value((_q_)))
 #define HG_IS_QINT(_v_)				\
